web_app/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
python-jose[cryptography]==3.5.0
python-dotenv==1.1.1
pydantic==2.11.7
livekit-api==1.0.6
numpy==1.26.4
scikit-learn==1.3.2
joblib==1.5.1
opencv-python==4.8.1.78
mediapipe==0.10.21

